Create Image loop back for workplaces PowerShell Script
I made this for a co worker to use on their bosses computer, the image is random cars then the message is to "Have a great shift!". This is a PowerShell script that will pull a random Image for a URL to create a pop up window on a loop, When the pop up is closed the loop will continue until the final pop up message is closed. The script will pull the image and save it to a temp folder then delete the image after words. I also added a bypass for the execution policy for PowerShell on Windows OS.